+++ infrastructure/cluster/provider.tf
@@ -4,6 +4,6 @@ provider "digitalocean" {
 
 /* Create a new SSH key */
 resource "digitalocean_ssh_key" "default" {
-    name = "${var.key_name}"
-    public_key = "${file(var.key_file)}"
+  name = "${var.key_name}"
+  public_key = "${file(var.key_file)}"
 }